fix: [김소연] 모든 사용자가 상품상세페이지의 리뷰를 클릭했을 때 해당 상품의 리뷰를 UI에 매핑하는 시스템[DIY-SALES-WEB-UI 69]
